This report, by its very length, defends itself against the risk of being read

+++ 0x05-pointers_arrays_strings/2-strlen.c
@@ -10,10 +10,10 @@
  * @str: The string to get the length of.
  *
  * Return: The length of @str.
- */
-size_t _strlen(const char *str)
+*/
+int _strlen(const char *str)
 {
-	size_t length = 0;
+	int length = 0;
 
 	while (*str++)
 		length++;
